Add formatting of message attachment text and field values
